Add stay delete exception when reservation is active from guest

+++ src/main/java/com/example/staybooking/service/StayService.java
@@ -1,17 +1,17 @@
 package com.example.staybooking.service;
 
+import com.example.staybooking.exception.StayDeleteException;
 import com.example.staybooking.exception.StayNotExistException;
-import com.example.staybooking.model.Location;
-import com.example.staybooking.model.Stay;
-import com.example.staybooking.model.StayImage;
-import com.example.staybooking.model.User;
+import com.example.staybooking.model.*;
 import com.example.staybooking.repository.LocationRepository;
+import com.example.staybooking.repository.ReservationRepository;
 import com.example.staybooking.repository.StayRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.awt.*;
+import java.time.LocalDate;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -24,13 +24,15 @@ public class StayService {
     private final StayRepository stayRepository;
     private final GeoCodingService geoCodingService;
     private final LocationRepository locationRepository;
+    private final ReservationRepository reservationRepository;
 
 
-    public StayService(ImageStorageService imageStorageService, StayRepository stayRepository, GeoCodingService geoCodingService, LocationRepository locationRepository) {
+    public StayService(ImageStorageService imageStorageService, StayRepository stayRepository, GeoCodingService geoCodingService, LocationRepository locationRepository, ReservationRepository reservationRepository) {
         this.imageStorageService = imageStorageService;
         this.stayRepository = stayRepository;
         this.geoCodingService = geoCodingService;
         this.locationRepository = locationRepository;
+        this.reservationRepository = reservationRepository;
     }
 
 
@@ -65,12 +67,19 @@ public class StayService {
     }
 
 
-    public void delete(Long stayId, String username) throws StayNotExistException {
+    public void delete(Long stayId, String username) throws StayNotExistException, StayDeleteException {
         Stay stay = stayRepository.findByIdAndHost(stayId, new User.Builder().setUsername(username).build());
         if (stay == null) {
             throw new StayNotExistException("Stay doesn't exist");
         }
+
+
+        List<Reservation> reservations = reservationRepository.findByStayAndCheckoutDateAfter(stay, LocalDate.now());
+        if (reservations != null && !reservations.isEmpty()) {
+            throw new StayDeleteException("Cannot delete stay with active reservation");
+        }
         stayRepository.deleteById(stayId);
     }
 }
+
